refactoring controllers
